cleanup on transaction commit
